dcp-t5(005/006): move factorytarget under chat_sessions and reconcile baselines
Relocate pkg/transports/mapping/factorytarget to
pkg/services/chat_sessions/transports/acp/factorytarget, creating
chat_sessions's first transports/ directory.

Filed under the acp protocol segment rather than mapping. The package doc
states it "projects a resolved Chat Sessions Factory target-catalog result
onto the merged ACP Factory target select-option contract, so ACP
transports consume one deterministic mapping"; it imports the ACP SDK and
pkg/transports/acp, and its only consumers are pkg/transports/acp/internal/
stdio. Its protocol is ACP, so acp/ is the truthful segment. It is also the
placement the boundary rule already sanctions: pkgboundarycheck's
isMatchingServiceOwnedTransportConsumer lets a protocol root consume a
service-owned adapter of that same protocol, so an ACP consumer of an acp/
adapter needs no exemption. Filing it under mapping/ instead produced two
"prohibited external service subpackage import" findings that could only be
cleared by widening or exempting that rule.

Baseline reconciliation for all five moved packages:
- go-unit / go-functional coverage minimums: entries renamed to the new
  import paths with measured values preserved, manifests re-sorted.
- backend-exemption-budget.json: the one factorytarget cyclomatic-complexity
  exemption is retargeted and moved to its new sorted position.
- service-construction-baseline.json: records the initsetup test's
  pre-existing providers/wire.NewService construction. The call and its line
  number are byte-identical before and after this lane; moving the file
  under pkg/services changed its fingerprint, so pkg-boundary's base-ref
  differential re-surfaced already-recorded debt. Entered with the same
  stage and deletionGate as its three siblings.
- backend-package-file-count.json is deliberately NOT touched: that baseline
  only admits packages exceeding 15 files and the five moved packages hold
  1-6, so the entry the PRD called for would fail validation twice.
- service-cycle-ceiling.json untouched; measured weight stays 30.

Owner registries in internal/ownershipinventory and
cmd/packagetargetmanifestcheck now list "transports" as an expected
chat_sessions child, matching the four other owners that already declare it.
