+++ docs/internal/baselines/service-construction-baseline.json
@@ -27,6 +27,15 @@
       "count": 1,
       "stage": "wire-injection-full-blow",
       "deletionGate": "inject the already-constructed service role from pkg/wire or move the invariant to the owning service"
+    },
+    {
+      "owner": "providers",
+      "importPath": "github.com/portpowered/infinite-you/pkg/services/providers/wire",
+      "symbol": "NewService",
+      "filePath": "pkg/services/operator_settings/transports/cli/initsetup/init_test.go",
+      "count": 1,
+      "stage": "wire-injection-full-blow",
+      "deletionGate": "inject the already-constructed service role from pkg/wire or move the invariant to the owning service"
     }
   ]
 }
