+++ pkg/services/chat_sessions/transports/acp/factorytarget/factory_target_option_test.go
@@ -6,7 +6,7 @@ import (
 
 	chatsessions "github.com/portpowered/infinite-you/pkg/services/chat_sessions"
 	"github.com/portpowered/infinite-you/pkg/transports/acp"
-	"github.com/portpowered/infinite-you/pkg/transports/mapping/factorytarget"
+	"github.com/portpowered/infinite-you/pkg/services/chat_sessions/transports/acp/factorytarget"
 )
 
 // pkgmaintcheck:ignore-cyclomatic-complexity pre-existing baseline debt recorded 2026-08-08; refactor this code below the maintainability threshold and remove this exemption
